fix layout of caption dropdowns & buttons
