+++ packages/datetime2/src/components/date-range-picker3/_date-range-picker3.scss
@@ -2,6 +2,8 @@
 // Licensed under the Apache License, Version 2.0.
 
 @import "@blueprintjs/colors/lib/scss/colors";
+@import "@blueprintjs/core/src/common/variables";
+@import "@blueprintjs/core/src/common/flex";
 @import "@blueprintjs/core/src/components/button/common";
 @import "@blueprintjs/datetime/src/common";
 
@@ -12,24 +14,60 @@
   // ensure min-widths are set correctly for variants of contiguous months, single month, and shortcuts
   &.#{$ns}-daterangepicker-contiguous .rdp {
     min-width: $datepicker-min-width + $pt-grid-size;
+
+    .rdp-caption {
+      // the default "caption" block + "nav" absolute position styles place the nav buttons slightly off,
+      // so we use flex instead
+      @include pt-flex-container(row);
+      // we can't inject a <Divider /> when only overriding the dropdown & icon components, so we add a CSS divider instead
+      border-bottom: solid 1px $pt-divider-black;
+      justify-content: space-between;
+      padding-bottom: $datepicker-padding;
+    }
+
+    .rdp-caption_start .rdp-caption {
+      flex-direction: row-reverse;
+
+      &::before {
+        // use a dummy pseudo element to achieve our desired flex box layout where we want the dropdowns to
+        // be centered over the calendar (even when there is an extra-wide footer element) while the nav buttons
+        // are justified on the corresponding start or end edge
+        content: "";
+        height: $pt-button-height;
+        width: $pt-button-height;
+      }
+    }
+
+    .rdp-caption_end .rdp-caption {
+      flex-direction: row;
+
+      &::before {
+        content: "";
+        height: $pt-button-height;
+        width: $pt-button-height;
+      }
+    }
+
+    .rdp-nav {
+      left: initial;
+      position: initial;
+      top: initial;
+      transform: none;
+    }
+
+    .#{$ns}-html-select select {
+      font-weight: 600;
+      padding-left: $datepicker-padding;
+      padding-right: $pt-icon-size-standard;
+
+      + .#{$ns}-icon {
+        right: 2px;
+      }
+    }
   }
 
   &.#{$ns}-daterangepicker-single-month .rdp {
     min-width: $datepicker-min-width;
-  }
-
-  .rdp-multiple_months {
-    .rdp-caption_dropdowns select {
-      font-weight: 600;
-    }
-
-    .rdp-caption_start .rdp-caption_dropdowns {
-      padding-left: $pt-button-height;
-    }
-
-    .rdp-caption_end .rdp-caption_dropdowns {
-      padding-right: $pt-button-height;
-    }
   }
 
   // need some extra specificity to override react-day-picker styles
