+++ packages/datetime2/src/components/react-day-picker/datePicker3Dropdown.tsx
@@ -19,6 +19,9 @@ import { DropdownProps } from "react-day-picker";
 
 import { HTMLSelect } from "@blueprintjs/core";
 
+import { Classes } from "../../classes";
+import { useMonthSelectRightOffset } from "../../common/useMonthSelectRightOffset";
+
 /**
  * Custom react-day-picker dropdown component which implements Blueprint's datepicker design
  * for month and year dropdowns.
@@ -26,9 +29,20 @@ import { HTMLSelect } from "@blueprintjs/core";
  * @see https://react-day-picker.js.org/guides/custom-components
  */
 export function DatePicker3Dropdown({ caption, children, ...props }: DropdownProps) {
+    const containerElement = React.useRef<HTMLDivElement>(null);
+    const selectElement = React.useRef<HTMLSelectElement>(null);
+
+    // Use a custom hook to adjust the position of the position of the HTMLSelect icon to appear right next to
+    // the month name. N.B. we expect props.caption to be a simple string representing the month name.
+    const displayedMonthText = typeof caption === "string" ? caption : "";
+    const monthSelectRightOffset = useMonthSelectRightOffset(selectElement, containerElement, displayedMonthText);
+    const iconProps = props.name === "months" ? { style: { right: monthSelectRightOffset } } : {};
+
     return (
-        <HTMLSelect minimal={true} {...props}>
-            {children}
-        </HTMLSelect>
+        <div className={Classes.DATEPICKER_DROPDOWN_CONTAINER} ref={containerElement}>
+            <HTMLSelect iconProps={iconProps} minimal={true} ref={selectElement} {...props}>
+                {children}
+            </HTMLSelect>
+        </div>
     );
 }
